fixing errors

+++ database/migrations/2024_02_27_182453_add_conductor_to_viaje_table.php
@@ -28,14 +28,11 @@ return new class extends Migration
      */
     public function down()
     {
-        // drop the keys
         Schema::table('viajes', function (Blueprint $table) {
-            $table->dropForeign('viaje_conductor_id_foreign');
-        });
-
-        // drop the actual columns
-        Schema::table('viajes', function (Blueprint $table) {
-            $table->dropColumn('conductor_id');
+            if (Schema::hasColumn('viajes', 'conductor_id')) {
+                $table->dropForeign(['conductor_id']);
+                $table->dropColumn('conductor_id');
+            }
         });
     }
 };
